Add analysis of later edits to files
This adds a notebook analyzing when edits to files occur relative
to when the file was created, for edits in 2019. We look at both
SDC and other types of edits, and identify 30 days as a reasonable
threshold for what "later" means. Then measures proportion of edits
by type (SDC and non-SDC edits).

Minor modifications to part 2 to also count bot edits.