Add notes on exploration of wbremoveclaims
This adds a text file with notes on the exploration of
the format of edit comments containing "wbremoveclaims".
We look at whether the edit comments mention specific properties,
if we can tell that a single property is removed or if the edit
removed multiple properties, and what the impact of the latter
might be.